Fixes needed 8/24/17
Running, but taking too long for merging. Looked at output and it's probably
because of way too many adducts being found (e.g. >800 for a dataset of
~1000 metabolites in OE_(C8-pos)). There are literal duplicates adducts
being found and also need to fix the M+1 issue (M+1 is a mapping on its
own). Finally, need to make sure the reverse mappings aren't being recorded
redundantly

Signed-off-by: gkreder <gkreder@gmail.com>